Another huge update, File input and output
Can now save a bot's dna using a string, this string can be saved as a text file, and to load the bots you can simply drag and drop the file into the editor scene and the bot will appear. This required some reworking of the DNA function toDataString to include the rotation of the joints. The DNA constructor is used to parse the string and build a bot. This makes the program run much more smoothly and enables more complex bots to be created and then saved for future use.